Fixed deleteById endpoint in customer service: now throws an error if customer to be deleted does not exist

+++ services/customer/src/main/java/com/dragand/ecommerce/service/CustomerService.java
@@ -73,6 +73,9 @@ public class CustomerService {
     }
 
     public void deleteCustomer(String customerId) {
+        if (!customerRepository.existsById(customerId)) {
+            throw new CustomerNotFoundException("Customer with ID " + customerId + " not found");
+        }
         customerRepository.deleteById(customerId);
     }
 }
